webapi/tct-calendar-tizen-tests/calendar/CalendarItem_convertToString.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.
Neither the name of Intel Corporation nor the names of its contributors
may be used to endorse or promote products derived from this work without
specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jenny Cao <jenny.q.cao@intel.com>
        Yanan Xu <yananx.xu@intel.com>
        Karol Surma <k.surma@samsung.com>
-->

<html>
<head>
<title>CalendarItem_convertToString</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/calendar_common.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: CalendarItem_convertToString
//==== LABEL Check if CalendarItem.convertToString method works properly
//==== SPEC Tizen Web API:Social:Calendar:CalendarItem:convertToString M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/calendar.html
//==== TEST_CRITERIA MMINA
test(function () {
    var eventInit, ev, eventStr;
    eventInit = {description:"HTML5 Introduction",
              summary:"HTML5 Webinar ",
              startDate: new tizen.TZDate(2011, 3, 30, 10, 0),
              duration: new tizen.TimeDuration(1, "HOURS"),
              location:"Huesca"};
    ev = new tizen.CalendarEvent(eventInit);
    eventStr = ev.convertToString("ICALENDAR_20");
    assert_not_equals(eventStr.length, 0, "Incorrect output from convertToString() method");
}, document.title);

</script>
</body>
</html>
